Thats the change i wanted

+++ src/lib/fetching.js
@@ -4,21 +4,28 @@ import { revalidatePath } from "next/cache";
 import { DJANGO_API_ENDPOINT } from "@/config/defaults";
 
 
-const API_CLIENTES_URL = "/api/clientes/";
-const API_UCS_URL = `$/api/ucs/`;
-const API_PROJETOS_URL = `/api/projetos/`;
+const DJANGO_API_CLIENTES_URL = `${DJANGO_API_ENDPOINT}/clientes/`;
+const DJANGO_API_UCS_URL = `${DJANGO_API_ENDPOINT}/ucs/`;
+const DJANGO_API_PROJETOS_URL = `${DJANGO_API_ENDPOINT}/projetos/`;
 const DJANGO_API_MODULOS_URL = `${DJANGO_API_ENDPOINT}/modulos/`;
 
 
 export async function fetchClientes() {
+    const authToken = getToken();
+    if (!authToken) {
+        return NextResponse.json({}, {status: 401});
+    }
 
     const options = { 
         method: "GET",
         headers: {
             "Content-Type": "application/json",
+            "Accept": "application/json",
+            "Authorization": `Bearer ${authToken}`
         }
     }
-    const response = await fetch(API_CLIENTES_URL, options);
+    const response = await fetch(DJANGO_API_CLIENTES_URL, options);
+    // console.log(response);
     const result = await response.json();
     const status = response.status;
 
@@ -26,11 +33,18 @@ export async function fetchClientes() {
 }
 
 export async function fetchCliente(id) {
-  const CLIENTE_DETAIL_URL = `${API_CLIENTES_URL}${id}/`
+  const CLIENTE_DETAIL_URL = `${DJANGO_API_CLIENTES_URL}${id}/`
+  const authToken = getToken();
+  if (!authToken) {
+    return NextResponse.json({}, {status: 401});
+  }
+
   const options = {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
     },
   }
   const response = await fetch(CLIENTE_DETAIL_URL, options);
@@ -42,12 +56,18 @@ export async function fetchCliente(id) {
 
 export async function deleteCliente(id) {
   "use server"
-  const DELETE_CLIENTE_API_URL = `${API_CLIENTES_URL}${id}/`;
+  const DELETE_CLIENTE_API_URL = `${DJANGO_API_CLIENTES_URL}${id}/`;
+  const authToken = getToken();
+  if (!authToken) {
+    return NextResponse.json({}, { status: 401 });
+  }
 
   const options = {
     method: "DELETE",
     headers: {
       "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
     },
   }
 
@@ -66,12 +86,18 @@ export async function deleteCliente(id) {
 
 export async function deleteUC(client_id, uc_id) {
   "use server"
-  const DELETE_UC_API_URL = `${API_UCS_URL}${client_id}/${uc_id}/`;
+  const DELETE_UC_API_URL = `${DJANGO_API_UCS_URL}${client_id}/${uc_id}/`;
+  const authToken = getToken();
+  if (!authToken) {
+    return NextResponse.json({}, { status: 401 });
+  }
 
   const options = {
     method: "DELETE",
     headers: {
       "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
     },
   }
 
@@ -88,13 +114,41 @@ export async function deleteUC(client_id, uc_id) {
   }
 }
 
-export async function fetchUCs(client_id) {
-  const UCS_LIST_URL = `${API_UCS_URL}${client_id}/`
+export async function fetchCategoria(categoria) {
+  const CATEGORIA_DETAIL_URL = `${DJANGO_API_UCS_URL}categoria/${categoria}/`
+  const authToken = getToken();
+  if (!authToken) {
+    return NextResponse.json({}, {status: 401});
+  }
 
   const options = {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
+    },
+  }
+  const response = await fetch(CATEGORIA_DETAIL_URL, options);
+  if (response.ok) {
+    return await response.json();
+  }
+  return NextResponse.json({}, {status: 400});
+}
+
+export async function fetchUCs(client_id) {
+  const UCS_LIST_URL = `${DJANGO_API_UCS_URL}${client_id}/`
+  const authToken = getToken();
+  if (!authToken) {
+    return NextResponse.json({}, {status: 401});
+  }
+
+  const options = {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
     },
   }
 
@@ -111,12 +165,18 @@ export async function fetchUCs(client_id) {
 
 export async function fetchUC(client_id, uc_id) {
   "use server"
-  const UC_RETRIEVE_URL = `${API_UCS_URL}${client_id}/${uc_id}/`;
+  const UC_RETRIEVE_URL = `${DJANGO_API_UCS_URL}${client_id}/${uc_id}/`;
+  const authToken = getToken();
+  if (!authToken) {
+    return NextResponse.json({}, {status: 401});
+  }
 
   const options = {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
     },
   }
 
@@ -130,22 +190,31 @@ export async function fetchUC(client_id, uc_id) {
 }
 
 export async function fetchProject(client_id) {
-  const API_PROJECT_URL = `${API_PROJETOS_URL}${client_id}/`;
+  const DJANGO_API_PROJECT_URL = `${DJANGO_API_PROJETOS_URL}${client_id}/`;
+  const authToken = getToken();
+
+  if (!authToken) {
+    return NextResponse.json({}, { status: 401 });
+  }
 
   const options = {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
+      "Accept": "application/json",
+      "Authorization": `Bearer ${authToken}`,
     },
   }
 
-  const response = await fetch(API_PROJECT_URL, options);
+  const response = await fetch(DJANGO_API_PROJECT_URL, options);
 
   if (response.ok) {
     return await response.json();
   }
 
   return NextResponse.json({}, {status: 400});
+
+
 }
 
 export async function getModulos() {
